Data fun
Using Linq I built a quick and dirty way to update the database from
computer to computer.

Added a connection string to web.config.
Added the start of NewBuild.
Added the Character model and some methods to load data into the
database and updated the gitignore so no App_Data files would try to be
added to git.